requirements.txt

asgiref==3.7.2
azure-ai-documentintelligence==1.0.0b2
azure-ai-formrecognizer==3.3.0
azure-common==1.1.28
azure-communication-identity==1.5.0
azure-core==1.30.1
azure-identity==1.16.0
Babel==2.14.0
billiard==4.2.0
boto3==1.34.131
botocore==1.34.131
cachetools==5.4.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.1.1
cron-descriptor==1.4.3
cryptography==42.0.5
cycler==0.12.1
Django==4.2.6
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-money==3.4.1
django-storages==1.14.3
django-timezone-field==7.0
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
fonttools==4.53.1
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.18.0
google-cloud-vision==3.7.4
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
httplib2==0.22.0
idna==3.7
isodate==0.6.1
jmespath==1.0.1
kiwisolver==1.4.5
kombu==5.3.7
lxml==4.9.4
matplotlib==3.7.5
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
numpy==1.24.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.0
pandas==2.0.3
pdf2image==1.17.0
pdfminer.six==20231228
pillow==10.3.0
portalocker==2.8.2
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==5.27.3
py-moneyed==3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
pytesseract==0.3.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.16
vine==5.1.0
wcwidth==0.2.13
